chore: login method moved to models/api/login
